types/shared_cos_lease: split FlowRrRing tests back + fix stale comment refs
Address Codex MERGE-NEEDS-MINOR on #1041:

1. The moved test mod in shared_cos_lease.rs included 7 FlowRrRing
   tests that exercise FlowRrRing (which stays in types/mod.rs).
   Split those tests out into a fresh `mod flow_rr_ring_tests` block
   at the end of types/mod.rs, colocated with FlowRrRing's struct +
   impl per modularity-discipline test-colocation.

2. Stale comment references to `types.rs` updated to point at the
   new directory module:
   - cos/queue_ops/pop.rs:66 (see types.rs) → (see types/mod.rs)
   - cos/admission.rs:31 invariants in types.rs → invariants in types/mod.rs
   - cos/queue_ops/v_min.rs:20 PaddedVtimeSlot::publish lives in
     types.rs → in types/shared_cos_lease.rs

build clean, 865 Rust tests pass.
